# openapi-formatted.json
{
    "openapi": "3.0.0",
    "paths": {
...
        "/api/v1/law-documents": {
            "post": {
                "operationId": "LawDocumentController_create",
                "summary": "\u521b\u5efa\u6cd5\u89c4\u4e0e\u6807\u51c6",
                "parameters": [],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreateLawDocumentDto"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": ""
                    }
                },
                "tags": [
                    "law-document"
                ]
            }
        },
        "/api/v1/law-documents/list": {
            "get": {
                "operationId": "LawDocumentController_findAll",
                "summary": "\u83b7\u53d6\u6cd5\u89c4\u4e0e\u6807\u51c6\u5217\u8868",
                "parameters": [
                    {
                        "name": "page",
                        "required": false,
                        "in": "query",
                        "description": "\u9875\u7801",
                        "schema": {
                            "default": 1,
                            "example": 1,
                            "type": "number"
                        }
                    },
                    {
                        "name": "pageSize",
                        "required": false,
                        "in": "query",
                        "description": "\u6bcf\u9875\u6570\u91cf",
                        "schema": {
                            "default": 10,
                            "example": 10,
                            "type": "number"
                        }
                    },
                    {
                        "name": "keyword",
                        "required": false,
                        "in": "query",
                        "description": "\u5173\u952e\u8bcd\u641c\u7d22",
                        "schema": {
                            "example": "\u5408\u540c",
                            "type": "string"
                        }
                    },
                    {
                        "name": "status",
                        "required": false,
                        "in": "query",
                        "description": "\u72b6\u6001\uff1a1-\u542f\u7528\uff0c0-\u505c\u7528",
                        "schema": {
                            "example": 1,
                            "enum": [
                                0,
                                1
                            ],
                            "type": "number"
                        }
                    },
                    {
                        "name": "lawCode",
                        "required": false,
                        "in": "query",
                        "description": "\u6cd5\u89c4\u7f16\u53f7",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "lawCategory",
                        "required": false,
                        "in": "query",
                        "description": "\u6587\u79cd\u7c7b\u522b",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "lawStatus",
                        "required": false,
                        "in": "query",
                        "description": "\u5f53\u524d\u72b6\u6001",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    }
                },
                "tags": [
                    "law-document"
                ]
            }
        },
...
        "/api/v1/law-documents/import": {
            "post": {
                "operationId": "LawDocumentController_importExcel",
                "summary": "Excel\u6279\u91cf\u5bfc\u5165",
                "parameters": [],
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "file": {
                                        "type": "string",
                                        "format": "binary"
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": ""
                    }
                },
                "tags": [
                    "law-document"
                ]
            }
        },
        "/api/v1/law-documents/{id}": {
            "get": {
                "operationId": "LawDocumentController_findOne",
                "summary": "\u83b7\u53d6\u6cd5\u89c4\u4e0e\u6807\u51c6\u8be6\u60c5",
                "parameters": [
                    {
                        "name": "id",
                        "required": true,
                        "in": "path",
                        "schema": {
                            "type": "number"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    }
                },
                "tags": [
                    "law-document"
                ]
            },
            "put": {
                "operationId": "LawDocumentController_update",
                "summary": "\u66f4\u65b0\u6cd5\u89c4\u4e0e\u6807\u51c6",
                "parameters": [
                    {
                        "name": "id",
                        "required": true,
                        "in": "path",
                        "schema": {
                            "type": "number"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateLawDocumentDto"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": ""
                    }
                },
                "tags": [
                    "law-document"
                ]
            },
            "delete": {
                "operationId": "LawDocumentController_remove",
                "summary": "\u5220\u9664/\u505c\u7528\u6cd5\u89c4\u4e0e\u6807\u51c6",
                "parameters": [
                    {
                        "name": "id",
                        "required": true,
                        "in": "path",
                        "schema": {
                            "type": "number"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    }
                },
                "tags": [
                    "law-document"
                ]
            }
        },
        "/api/v1/law-documents/batch/enable": {
            "post": {
                "operationId": "LawDocumentController_batchEnable",
                "summary": "\u6279\u91cf\u542f\u7528",
                "parameters": [],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/BatchIdsDto"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": ""
                    }
                },
                "tags": [
                    "law-document"
                ]
            }
        },
        "/api/v1/law-documents/batch/disable": {
            "post": {
                "operationId": "LawDocumentController_batchDisable",
                "summary": "\u6279\u91cf\u505c\u7528",
                "parameters": [],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/BatchIdsDto"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": ""
                    }
                },
                "tags": [
                    "law-document"
                ]
            }
        },
        "/api/v1/law-documents/batch/delete": {
            "post": {
                "operationId": "LawDocumentController_batchDelete",
                "summary": "\u6279\u91cf\u5220\u9664",
                "parameters": [],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/BatchIdsDto"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": ""
                    }
                },
                "tags": [
                    "law-document"
                ]
            }
        },
...
    },
    "info": {
        "title": "\u5de5\u7a0b\u54a8\u8be2\u5168\u4e1a\u52a1\u6570\u636e\u5e93\u5e73\u53f0 API",
        "description": "\n## \u6982\u8ff0\n\u5de5\u7a0b\u54a8\u8be2\u5168\u4e1a\u52a1\u6570\u636e\u5e93\u5e73\u53f0\u540e\u7aef\u63a5\u53e3\u6587\u6863 v1.0\n\n## \u8ba4\u8bc1\n\u6240\u6709\u63a5\u53e3\uff08\u9664\u767b\u5f55\u3001\u6ce8\u518c\u3001\u5065\u5eb7\u68c0\u67e5\u5916\uff09\u5747\u9700\u8981 JWT \u8ba4\u8bc1\u3002\n\u8bf7\u5728\u8bf7\u6c42\u5934\u4e2d\u6dfb\u52a0\uff1a`Authorization: Bearer <token>`\n\n## \u54cd\u5e94\u683c\u5f0f\n- \u6210\u529f\uff1a`{ data: T, meta?: { page, pageSize, total, totalPages } }`\n- \u5931\u8d25\uff1a`{ code: string, message: string, traceId: string, details?: any }`\n\n## \u5206\u9875\n- \u9ed8\u8ba4 page=1, pageSize=10\n- \u6700\u5927 pageSize=1000\n    ",
        "version": "1.0",
        "contact": {}
    },
...
    "servers": [],
    "components": {
        "securitySchemes": {
            "JWT": {
                "scheme": "bearer",
                "bearerFormat": "JWT",
                "type": "http",
                "description": "\u8f93\u5165 JWT Token \u8fdb\u884c\u8ba4\u8bc1"
            }
        },
        "schemas": {
...
            "BatchIdsDto": {
                "type": "object",
                "properties": {
                    "ids": {
                        "description": "ID\u6570\u7ec4",
                        "type": "array",
                        "items": {
                            "type": "number"
                        }
                    }
                },
                "required": [
                    "ids"
                ]
            },
...
            "CreateLawDocumentDto": {
                "type": "object",
                "properties": {
                    "lawCode": {
                        "type": "string",
                        "description": "\u6cd5\u89c4\u7f16\u53f7",
                        "example": "GB50500-2013"
                    },
                    "lawName": {
                        "type": "string",
                        "description": "\u6cd5\u89c4\u540d\u79f0",
                        "example": "\u5efa\u8bbe\u5de5\u7a0b\u5de5\u7a0b\u91cf\u6e05\u5355\u8ba1\u4ef7\u89c4\u8303"
                    },
                    "lawCategory": {
                        "type": "string",
                        "description": "\u6587\u79cd\u7c7b\u522b",
                        "example": "\u56fd\u5bb6\u6807\u51c6"
                    },
                    "issueOrg": {
                        "type": "string",
                        "description": "\u53d1\u5e03\u5355\u4f4d",
                        "example": "\u4f4f\u623f\u548c\u57ce\u4e61\u5efa\u8bbe\u90e8"
                    },
                    "issueDate": {
                        "type": "string",
                        "description": "\u53d1\u5e03\u65e5\u671f",
                        "example": "2013-04-01"
                    },
                    "effectiveDate": {
                        "type": "string",
                        "description": "\u5b9e\u65bd\u65e5\u671f",
                        "example": "2013-07-01"
                    },
                    "expiryDate": {
                        "type": "string",
                        "description": "\u5931\u6548\u65e5\u671f",
                        "example": null
                    },
                    "regionScope": {
                        "type": "string",
                        "description": "\u9002\u7528\u5730\u533a\u8303\u56f4",
                        "example": "\u5168\u56fd"
                    },
                    "industryScope": {
                        "type": "string",
                        "description": "\u9002\u7528\u884c\u4e1a\u8303\u56f4",
                        "example": "\u5efa\u8bbe\u5de5\u7a0b"
                    },
                    "lawStatus": {
                        "type": "string",
                        "description": "\u5f53\u524d\u72b6\u6001",
                        "example": "\u73b0\u884c"
                    },
                    "filePath": {
                        "type": "string",
                        "description": "\u539f\u6587\u6587\u4ef6\u4f4d\u7f6e",
                        "example": "/uploads/laws/GB50500-2013.pdf"
                    },
                    "summary": {
                        "type": "string",
                        "description": "\u6458\u8981/\u8981\u70b9\u8bf4\u660e",
                        "example": "\u89c4\u5b9a\u4e86\u5efa\u8bbe\u5de5\u7a0b\u5de5\u7a0b\u91cf\u6e05\u5355\u7684\u7f16\u5236\u65b9\u6cd5\u548c\u8ba1\u4ef7\u89c4\u5219"
                    },
                    "remark": {
                        "type": "string",
                        "description": "\u5907\u6ce8",
                        "example": "\u66ff\u4ee3GB50500-2008"
                    }
                },
                "required": [
                    "lawCode",
                    "lawName"
                ]
            },
            "UpdateLawDocumentDto": {
                "type": "object",
                "properties": {
                    "lawCode": {
                        "type": "string",
                        "description": "\u6cd5\u89c4\u7f16\u53f7",
                        "example": "GB50500-2013"
                    },
                    "lawName": {
                        "type": "string",
                        "description": "\u6cd5\u89c4\u540d\u79f0",
                        "example": "\u5efa\u8bbe\u5de5\u7a0b\u5de5\u7a0b\u91cf\u6e05\u5355\u8ba1\u4ef7\u89c4\u8303"
                    },
                    "lawCategory": {
                        "type": "string",
                        "description": "\u6587\u79cd\u7c7b\u522b",
                        "example": "\u56fd\u5bb6\u6807\u51c6"
                    },
                    "issueOrg": {
                        "type": "string",
                        "description": "\u53d1\u5e03\u5355\u4f4d",
                        "example": "\u4f4f\u623f\u548c\u57ce\u4e61\u5efa\u8bbe\u90e8"
                    },
                    "issueDate": {
                        "type": "string",
                        "description": "\u53d1\u5e03\u65e5\u671f",
                        "example": "2013-04-01"
                    },
                    "effectiveDate": {
                        "type": "string",
                        "description": "\u5b9e\u65bd\u65e5\u671f",
                        "example": "2013-07-01"
                    },
                    "expiryDate": {
                        "type": "string",
                        "description": "\u5931\u6548\u65e5\u671f",
                        "example": null
                    },
                    "regionScope": {
                        "type": "string",
                        "description": "\u9002\u7528\u5730\u533a\u8303\u56f4",
                        "example": "\u5168\u56fd"
                    },
                    "industryScope": {
                        "type": "string",
                        "description": "\u9002\u7528\u884c\u4e1a\u8303\u56f4",
                        "example": "\u5efa\u8bbe\u5de5\u7a0b"
                    },
                    "lawStatus": {
                        "type": "string",
                        "description": "\u5f53\u524d\u72b6\u6001",
                        "example": "\u73b0\u884c"
                    },
                    "filePath": {
                        "type": "string",
                        "description": "\u539f\u6587\u6587\u4ef6\u4f4d\u7f6e",
                        "example": "/uploads/laws/GB50500-2013.pdf"
                    },
                    "summary": {
                        "type": "string",
                        "description": "\u6458\u8981/\u8981\u70b9\u8bf4\u660e",
                        "example": "\u89c4\u5b9a\u4e86\u5efa\u8bbe\u5de5\u7a0b\u5de5\u7a0b\u91cf\u6e05\u5355\u7684\u7f16\u5236\u65b9\u6cd5\u548c\u8ba1\u4ef7\u89c4\u5219"
                    },
                    "remark": {
                        "type": "string",
                        "description": "\u5907\u6ce8",
                        "example": "\u66ff\u4ee3GB50500-2008"
                    },
                    "status": {
                        "type": "number",
                        "description": "\u72b6\u6001\uff1a1-\u542f\u7528\uff0c0-\u505c\u7528"
                    }
                }
            }
        }
    }
}